native_agent/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.23,<1
pydantic>=2.5,<3
python-dotenv>=1.0,<2
openai>=1.30,<2
typer>=0.12,<1
httpx>=0.27,<1
pytest>=8,<9
